Quartets
Problem

A partial split of a set S of n taxa models a partial character and is denoted by A∣B, where A and B are still the two disjoint subsets of taxa divided by the character. Unlike in the case of splits, we do not necessarily require that A∪B=S; (A∪B)c corresponds to those taxa for which we lack conclusive evidence regarding the character.

We can assemble a collection of partial characters into a generalized partial character table C in which the symbol x is placed in Ci,j if we do not have conclusive evidence regarding the jth taxon with respect to the ith partial character.

A quartet is a partial split A∣B in which both A and B contain precisely two elements. For the sake of simplicity, we often will consider quartets instead of partial characters. We say that a quartet A∣B is inferred from a partial split C∣D if A⊆C and B⊆D (or equivalently A⊆D and B⊆C). For example, {1,3}∣{2,4} and {3,5}∣{2,4} can be inferred from {1,3,5}∣{2,4}.

Given: A partial character table C.

Return: The collection of all quartets that can be inferred from the splits corresponding to the underlying characters of C.